+++ spec/support/controllers/params/empty_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Params
+  class EmptyController < ActionController::Base
+    include Pastore::Params
+  end
+end