+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 ## [Unreleased]
 
-## [0.1.0] - 2023-01-26
+## [0.1.0] - 2023-02-13
 
-- Initial release
+- [NEW] Implemented `Pastore::Params` feature for params validation.
+
+## [0.0.4] - 2023-02-02
+
+- [FIX] Correctly inherit default guards strategy from parent class.
+
+## [0.0.3] - 2023-02-02
+
+- [FIX] Correctly handle guards callback inheritance.
+
+## [0.0.2] - 2023-02-01
+
+- [NEW] Added some RSpec tests.
+- [NEW] Setup automated tests.
+
+## [0.0.1] - 2023-01-31
+
+- [NEW] First implementation of `Pastore::Guards` feature.